backend/requirements.txt
Django==4.2.7
django-environ==0.10.0
django-storages==1.13.2
djangorestframework==3.14.0

boto3==1.28.27
psycopg2-binary==2.9.7
sentry-sdk==1.29.2
stripe==7.1.0
django-q==1.3.9
pillow==10.1.0
